refactor(console): split the module catalog out of types.ts
types.ts was a 1824-line god file: wire types, builtin commands, nav
shapes, the whole MODULE_CATALOG (~1085 lines) and four domains' page
shapes in one declaration surface. The catalog now lives one file per
module under shared/lib/catalog/ (the extension story mirrors sesame's
all.go: a module is a file plus one line in catalog/index.ts), the
module-def vocabulary in catalog/module-def.ts, and the Govee /
ChannelPoints / Timers / Loyalty page shapes moved next to their
stores. MOD joins it: module ids were hardcoded per store, where a
typo compiled fine and silently missed the module blob.

Pure code motion — every comment travels verbatim; '@bagel/shared'
resolves exactly as before via the types.ts compat barrel.
